Refactored deprecated save and open panel code in ScriptExec, int-->NSInteger

+++ ScriptExec/ScriptExecController.h
@@ -103,8 +103,8 @@
     NSColor *textBackground;
     int textEncoding;
     
-    int execStyle;
-    int outputType;
+    NSInteger execStyle;
+    NSInteger outputType;
     BOOL isDroppable;
     BOOL remainRunning;
     BOOL secureScript;
